Extract `Type1Parser` from fonts.js
